# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mixd",
    "description": "Personal music metadata hub",
    "version": "0.10.0"
  },
  "paths": {
...
    "/api/v1/chat/feedback": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Post Chat Feedback",
        "operationId": "post_chat_feedback_api_v1_chat_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Post Chat",
        "operationId": "post_chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatFeedbackRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 20480,
            "title": "Prompt"
          },
          "generated_workflow_def": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue-Input"
            },
            "type": "object",
            "title": "Generated Workflow Def"
          },
          "signal": {
            "type": "string",
            "enum": [
              "positive",
              "negative"
            ],
            "title": "Signal"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "generated_workflow_def",
          "signal"
        ],
        "title": "ChatFeedbackRequest",
        "description": "Thumbs up/down on a generated workflow, with the triggering prompt."
      },
      "ChatFeedbackResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ChatFeedbackResponse",
        "description": "The persisted feedback row's id."
      },
      "ChatMessageInput": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 20480,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessageInput",
        "description": "One conversation turn from the client."
      },
      "ChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageInput"
            },
            "type": "array",
            "title": "Messages"
          },
          "confirmation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfirmationInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Date"
          },
          "effort": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EffortLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_workflow_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Workflow Id"
          },
          "page": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatRequest",
        "description": "A chat request: message history plus optional confirmation and effort."
      },
...
      "ConfirmationInput": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "action_id",
          "approved"
        ],
        "title": "ConfirmationInput",
        "description": "A user's approve/cancel decision on a proposed mutation."
      },
...
      "EffortLevel": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "xhigh",
          "max"
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue-Input": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue-Input"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue-Input"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}